+++ lib/startback/audit.rb
@@ -1,2 +1,3 @@
+require_relative 'audit/shared'
 require_relative 'audit/trailer'
 require_relative 'audit/prometheus'